+++ src/commands/diffWithNext.ts
@@ -3,7 +3,7 @@ import { Iterables } from '../system';
 import { commands, Range, TextDocumentShowOptions, TextEditor, Uri, window } from 'vscode';
 import { ActiveEditorCommand, Commands, getCommandUri } from './common';
 import { DiffWithCommandArgs } from './diffWith';
-import { GitLogCommit, GitService, GitUri } from '../gitService';
+import { GitLogCommit, GitService, GitStatusFile, GitUri } from '../gitService';
 import { Logger } from '../logger';
 import { Messages } from '../messages';
 
@@ -32,21 +32,26 @@ export class DiffWithNextCommand extends ActiveEditorCommand {
             args.line = editor === undefined ? 0 : editor.selection.active.line;
         }
 
+        const gitUri = await GitUri.fromUri(uri, this.git);
+        let status: GitStatusFile | undefined;
+
         if (args.commit === undefined || !(args.commit instanceof GitLogCommit) || args.range !== undefined) {
-            const gitUri = await GitUri.fromUri(uri, this.git);
-
             try {
-                // If the sha is missing or the file is uncommitted, treat it as a DiffWithWorking
-                if (gitUri.sha === undefined && await this.git.isFileUncommitted(gitUri)) {
-                    return commands.executeCommand(Commands.DiffWithWorking, uri);
-                }
-
                 const sha = args.commit === undefined ? gitUri.sha : args.commit.sha;
+
+                // If we are a fake "staged" sha, treat it as a DiffWithWorking
+                if (GitService.isStagedUncommitted(sha!)) return commands.executeCommand(Commands.DiffWithWorking, uri);
 
                 const log = await this.git.getLogForFile(gitUri.repoPath, gitUri.fsPath, undefined, { maxCount: sha !== undefined ? undefined : 2, range: args.range! });
                 if (log === undefined) return Messages.showFileNotUnderSourceControlWarningMessage('Unable to open compare');
 
                 args.commit = (sha && log.commits.get(sha)) || Iterables.first(log.commits.values());
+
+                // If the sha is missing or the file is uncommitted, treat it as a DiffWithWorking
+                if (gitUri.sha === undefined) {
+                    status = await this.git.getStatusForFile(gitUri.repoPath!, gitUri.fsPath);
+                    if (status !== undefined) return commands.executeCommand(Commands.DiffWithWorking, uri);
+                }
             }
             catch (ex) {
                 Logger.error(ex, 'DiffWithNextCommand', `getLogForFile(${gitUri.repoPath}, ${gitUri.fsPath})`);
@@ -54,7 +59,29 @@ export class DiffWithNextCommand extends ActiveEditorCommand {
             }
         }
 
-        if (args.commit.nextSha === undefined) return commands.executeCommand(Commands.DiffWithWorking, uri);
+        if (args.commit.nextSha === undefined) {
+            // Check if the file is staged
+            status = status || await this.git.getStatusForFile(gitUri.repoPath!, gitUri.fsPath);
+            if (status !== undefined && status.indexStatus === 'M') {
+                const diffArgs: DiffWithCommandArgs = {
+                    repoPath: args.commit.repoPath,
+                    lhs: {
+                        sha: args.commit.sha,
+                        uri: args.commit.uri
+                    },
+                    rhs: {
+                        sha: GitService.stagedUncommittedSha,
+                        uri: args.commit.uri
+                    },
+                    line: args.line,
+                    showOptions: args.showOptions
+                };
+
+                return commands.executeCommand(Commands.DiffWith, diffArgs);
+            }
+
+            return commands.executeCommand(Commands.DiffWithWorking, uri);
+        }
 
         const diffArgs: DiffWithCommandArgs = {
             repoPath: args.commit.repoPath,