+++ src/commands/diffWithPrevious.ts
@@ -37,8 +37,15 @@ export class DiffWithPreviousCommand extends ActiveEditorCommand {
             const gitUri = await GitUri.fromUri(uri, this.git);
 
             try {
-                const sha = args.commit === undefined ? gitUri.sha : args.commit.sha;
+                let sha = args.commit === undefined ? gitUri.sha : args.commit.sha;
                 if (sha === GitService.deletedSha) return Messages.showCommitHasNoPreviousCommitWarningMessage();
+
+                // If we are a fake "staged" sha, remove it
+                let isStagedUncommitted = false;
+                if (GitService.isStagedUncommitted(sha!)) {
+                    gitUri.sha = sha = undefined;
+                    isStagedUncommitted = true;
+                }
 
                 const log = await this.git.getLogForFile(gitUri.repoPath, gitUri.fsPath, sha, { maxCount: 2, range: args.range!, skipMerges: true });
                 if (log === undefined) return Messages.showFileNotUnderSourceControlWarningMessage('Unable to open compare');
@@ -46,8 +53,47 @@ export class DiffWithPreviousCommand extends ActiveEditorCommand {
                 args.commit = (sha && log.commits.get(sha)) || Iterables.first(log.commits.values());
 
                 // If the sha is missing and the file is uncommitted, then treat it as a DiffWithWorking
-                if (gitUri.sha === undefined && await this.git.isFileUncommitted(gitUri)) {
-                    return commands.executeCommand(Commands.DiffWithWorking, uri, { commit: args.commit, showOptions: args.showOptions } as DiffWithWorkingCommandArgs);
+                if (gitUri.sha === undefined) {
+                    const status = await this.git.getStatusForFile(gitUri.repoPath!, gitUri.fsPath);
+                    if (status !== undefined) {
+                        if (isStagedUncommitted) {
+                            const diffArgs: DiffWithCommandArgs = {
+                                repoPath: args.commit.repoPath,
+                                lhs: {
+                                    sha: args.commit.sha,
+                                    uri: args.commit.uri
+                                },
+                                rhs: {
+                                    sha: GitService.stagedUncommittedSha,
+                                    uri: args.commit.uri
+                                },
+                                line: args.line,
+                                showOptions: args.showOptions
+                            };
+                            return commands.executeCommand(Commands.DiffWith, diffArgs);
+                        }
+
+                        // Check if the file is staged
+                        if (status.indexStatus !== undefined) {
+                            const diffArgs: DiffWithCommandArgs = {
+                                repoPath: args.commit.repoPath,
+                                lhs: {
+                                    sha: GitService.stagedUncommittedSha,
+                                    uri: args.commit.uri
+                                },
+                                rhs: {
+                                    sha: '',
+                                    uri: args.commit.uri
+                                },
+                                line: args.line,
+                                showOptions: args.showOptions
+                            };
+
+                            return commands.executeCommand(Commands.DiffWith, diffArgs);
+                        }
+
+                        return commands.executeCommand(Commands.DiffWithWorking, uri, { commit: args.commit, showOptions: args.showOptions } as DiffWithWorkingCommandArgs);
+                    }
                 }
             }
             catch (ex) {