+++ src/views/fileHistoryNode.ts
@@ -3,7 +3,7 @@ import { Iterables } from '../system';
 import { Disposable, TreeItem, TreeItemCollapsibleState } from 'vscode';
 import { CommitFileNode, CommitFileNodeDisplayAs } from './commitFileNode';
 import { ExplorerNode, MessageNode, ResourceType } from './explorerNode';
-import { GitUri, Repository, RepositoryChange, RepositoryChangeEvent } from '../gitService';
+import { BlameabilityChangeEvent, BlameabilityChangeReason, GitCommitType, GitLogCommit, GitService, GitUri, Repository, RepositoryChange, RepositoryChangeEvent } from '../gitService';
 import { GitExplorer } from './gitExplorer';
 import { Logger } from '../logger';
 
@@ -20,10 +20,49 @@ export class FileHistoryNode extends ExplorerNode {
     async getChildren(): Promise<ExplorerNode[]> {
         this.updateSubscription();
 
-        const log = await this.explorer.git.getLogForFile(this.uri.repoPath, this.uri.fsPath, this.uri.sha);
-        if (log === undefined) return [new MessageNode('No file history')];
+        const children: ExplorerNode[] = [];
 
-        return [...Iterables.map(log.commits.values(), c => new CommitFileNode(c.fileStatuses[0], c, this.explorer, CommitFileNodeDisplayAs.CommitLabel | CommitFileNodeDisplayAs.StatusIcon))];
+        const status = await this.explorer.git.getStatusForFile(this.uri.repoPath!, this.uri.fsPath);
+        if (status !== undefined && (status.indexStatus !== undefined || status.workTreeStatus !== undefined)) {
+            let sha;
+            let previousSha;
+            if (status.workTreeStatus !== undefined) {
+                sha = GitService.uncommittedSha;
+                if (status.indexStatus !== undefined) {
+                    previousSha = GitService.stagedUncommittedSha;
+                }
+                else if (status.workTreeStatus !== '?') {
+                    previousSha = 'HEAD';
+                }
+            }
+            else {
+                sha = GitService.stagedUncommittedSha;
+                previousSha = 'HEAD';
+            }
+
+            const commit = new GitLogCommit(
+                GitCommitType.File,
+                this.uri.repoPath!,
+                sha,
+                status.fileName,
+                'You',
+                new Date(),
+                '',
+                status.status,
+                [status],
+                status.originalFileName,
+                previousSha,
+                status.originalFileName || status.fileName);
+            children.push(new CommitFileNode(status, commit, this.explorer, CommitFileNodeDisplayAs.CommitLabel | CommitFileNodeDisplayAs.StatusIcon));
+        }
+
+        const log = await this.explorer.git.getLogForFile(this.uri.repoPath, this.uri.fsPath, this.uri.sha);
+        if (log !== undefined) {
+            children.push(...Iterables.map(log.commits.values(), c => new CommitFileNode(c.fileStatuses[0], c, this.explorer, CommitFileNodeDisplayAs.CommitLabel | CommitFileNodeDisplayAs.StatusIcon)));
+        }
+
+        if (children.length === 0) return [new MessageNode('No file history')];
+        return children;
     }
 
     getTreeItem(): TreeItem {
@@ -45,7 +84,8 @@ export class FileHistoryNode extends ExplorerNode {
         if (this.explorer.autoRefresh) {
             this.disposable = this.disposable || Disposable.from(
                 this.explorer.onDidChangeAutoRefresh(this.onAutoRefreshChanged, this),
-                this.repo.onDidChange(this.onRepoChanged, this)
+                this.repo.onDidChange(this.onRepoChanged, this),
+                this.explorer.gitContextTracker.onDidChangeBlameability(this.onBlameabilityChanged, this)
             );
         }
         else if (this.disposable !== undefined) {
@@ -56,6 +96,14 @@ export class FileHistoryNode extends ExplorerNode {
 
     private onAutoRefreshChanged() {
         this.updateSubscription();
+    }
+
+    private onBlameabilityChanged(e: BlameabilityChangeEvent) {
+        if (!e.blameable || e.reason !== BlameabilityChangeReason.DocumentChanged) return;
+
+        // Logger.log(`RepositoryNode.onBlameabilityChanged(${e.reason}); triggering node refresh`);
+
+        this.explorer.refreshNode(this);
     }
 
     private onRepoChanged(e: RepositoryChangeEvent) {